Refactored fetchApi into two functions with a callback
We now have apiCall() with callback sendRiotTemplate().
This seperation  should make them easier to combine apiCall with other tempatate types (jade, react, etc).
It seems to work  but I can see it is not quite right.
Todos
Write tests
Pass in the name of the callBack rather than just hardcode it in.
Call the functions in the route file rather than the helper file
Might use async