Test if Post Exists Or not in api
Wrote two tests to see if a post exists or does not exist in the api.  This required more reworking of the functions in fetchApi.js and the way they are called in PostRoute.js.
Testing if a post exists is useful when we create the slug.  If there is already a post with the same slug we are creating we can amend it accordingly, perhaps versioning it with a number on the end (as WordPress does). ]